Move check_url_availability() to external.utils.functions.
