+++ components/server/src/external/routes/source.py
@@ -11,8 +11,10 @@ from model.actions import copy_source, move_item
 from model.data import MetricData, SourceData
 from model.queries import is_password_parameter
 from model.transformations import change_source_parameter
-from server_utilities.functions import check_url_availability, uuid
+from server_utilities.functions import uuid
 from server_utilities.type import EditScope, MetricId, ReportId, SourceId, SubjectId
+
+from ..utils.functions import check_url_availability
 
 from .plugins.auth_plugin import EDIT_REPORT_PERMISSION
