+++ components/server/src/server_utilities/functions.py
@@ -1,7 +1,6 @@
 """Utility functions."""
 
 import hashlib
-import re
 import uuid as _uuid
 from base64 import b64decode, b64encode
 from collections.abc import Callable, Hashable, Iterable, Iterator
@@ -9,14 +8,13 @@ from datetime import datetime, timezone
 from decimal import ROUND_HALF_UP, Decimal
 from typing import Tuple, TypeVar, cast
 
-import requests
 from cryptography.hazmat.backends import default_backend, openssl
 from cryptography.hazmat.primitives import serialization
 from cryptography.hazmat.primitives import hashes
 from cryptography.hazmat.primitives.asymmetric import padding
 from cryptography.fernet import Fernet
 
-from server_utilities.type import Direction, ReportId, URL
+from server_utilities.type import Direction, ReportId
 
 
 class DecryptionError(Exception):
@@ -60,35 +58,6 @@ def percentage(numerator: int, denominator: int, direction: Direction) -> int:
     if denominator == 0:
         return 0 if direction == "<" else 100
     return int((100 * Decimal(numerator) / Decimal(denominator)).to_integral_value(ROUND_HALF_UP))
-
-
-def check_url_availability(url: URL, source_parameters: dict[str, str]) -> dict[str, int | str]:
-    """Check the availability of the URL."""
-    credentials = _basic_auth_credentials(source_parameters)
-    headers = _headers(source_parameters)
-    try:
-        response = requests.get(url, auth=credentials, headers=headers, verify=False)  # noqa: DUO123, # nosec
-        return dict(status_code=response.status_code, reason=response.reason)
-    except Exception as exception_instance:  # pylint: disable=broad-except
-        exception_reason = str(exception_instance) or exception_instance.__class__.__name__
-        # If the reason contains an errno, only return the errno and accompanying text, and leave out the traceback
-        # that led to the error:
-        exception_reason = re.sub(r".*(\[errno \-?\d+\] [^\)^']+).*", r"\1", exception_reason, flags=re.IGNORECASE)
-        return dict(status_code=-1, reason=exception_reason)
-
-
-def _basic_auth_credentials(source_parameters) -> tuple[str, str] | None:
-    """Return the basic authentication credentials, if any."""
-    if private_token := source_parameters.get("private_token", ""):
-        return private_token, ""
-    username = source_parameters.get("username", "")
-    password = source_parameters.get("password", "")
-    return (username, password) if username or password else None
-
-
-def _headers(source_parameters) -> dict:
-    """Return the headers for the url-check."""
-    return {"Private-Token": source_parameters["private_token"]} if "private_token" in source_parameters else {}
 
 
 def symmetric_encrypt(message: bytes) -> tuple[bytes, bytes]:
